✅ test(workflows,workflows-hono): mixed export of agent signal types for workflow
