+++ src/server/workflows-hono/agent-signal/index.ts
@@ -1,8 +1,8 @@
 import { serve } from '@upstash/workflow/hono';
 import { Hono } from 'hono';
 
-import type { AgentSignalWorkflowRunPayload } from '@/server/workflows/agentSignal';
 import { runAgentSignalWorkflow } from '@/server/workflows/agentSignal/run';
+import type { AgentSignalWorkflowRunPayload } from '@/server/workflows/agentSignal/types';
 
 import { createWorkflowQstashClient } from '../qstashClient';
